pyproject.toml
[project]
name = "pip-bot"
version = "0.3.0"
description = "Personal Discord bot"
authors = [
    {name = "jorgebc", email = "jorge.buffa.casalis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.7.1,<3.0.0",
    "python-dotenv>=1.2.2,<2.0.0",
    "psutil>=7.2.2,<8.0.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "bot"}, {include = "cogs"}, {include = "config"}, {include = "services"}, {include = "utils"}]

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "ruff>=0.15.7,<0.16.0"
]
